+++ tests/support/stub/SiteController.php
@@ -167,7 +167,7 @@ final class SiteController extends Controller
         fwrite($tmpFile, 'This is a test file content.');
         rewind($tmpFile);
 
-        $tmpFilePath = stream_get_meta_data($tmpFile)['uri'];
+        $tmpFilePath = stream_get_meta_data($tmpFile)['uri'] ?? '';
 
         return $this->response->sendFile($tmpFilePath, 'testfile.txt', ['mimeType' => 'text/plain']);
     }