+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Bug #183: Remove redundant cookie tests from `ResponseAdapterTest` for clarity (@terabytesoftw)
 - Bug #184: Remove redundant remote IP and server port from `ServerParamsPsr7Test` for clarity (@terabytesoftw)
 - Bug #185: Consolidate `paths-ignore` configuration across workflow files for consistency, add logos for badges in `README.md` (@terabytesoftw)
+- Bug #186: Remove redundant test cases from `ServerRequestAdapterTest` for clarity (@terabytesoftw)
 
 ## 0.1.0 September 26, 2025